requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.6
pydantic==2.7.4
pydantic-settings==2.3.3
passlib==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
alembic==1.13.1

# Email dependencies (production-grade SendGrid)
sendgrid==6.11.0
jinja2==3.1.4

# Legacy SMTP dependencies (deprecated)
# aiosmtplib==3.0.1

# Development and code quality tools
pre-commit
black
flake8
isort
mypy
bandit
pytest

types-redis
types-requests
